+++ src/containerapp/azext_containerapp/custom.py
@@ -34,7 +34,7 @@ from knack.prompting import prompt_y_n, prompt as prompt_str
 from msrestazure.tools import parse_resource_id, is_valid_resource_id
 from msrest.exceptions import DeserializationError
 
-
+from .containerapp_decorator import ContainerAppCreateDecorator, BaseContainerAppDecorator
 from ._client_factory import handle_raw_exception, handle_non_404_exception
 from ._clients import ManagedEnvironmentClient, ContainerAppClient, GitHubActionClient, DaprComponentClient, StorageClient, AuthClient, WorkloadProfileClient, ContainerAppsJobClient
 from ._dev_service_utils import DevServiceUtils
@@ -92,7 +92,8 @@ from ._utils import (_validate_subscription_registered, _ensure_location_allowed
                      check_managed_cert_name_availability, prepare_managed_certificate_envelop,
                      get_default_workload_profile_name_from_env, get_default_workload_profiles, ensure_workload_profile_supported, _generate_secret_volume_name,
                      parse_service_bindings, get_linker_client, check_unique_bindings,
-                     get_current_mariner_tags, patchable_check, get_pack_exec_path, is_docker_running, AppType)
+                     get_current_mariner_tags, patchable_check, get_pack_exec_path, is_docker_running, trigger_workflow, AppType,
+                     format_location)
 from ._validators import validate_create, validate_revision_suffix
 from ._ssh_utils import (SSH_DEFAULT_ENCODING, WebSocketConnection, read_ssh, get_stdin_writer, SSH_CTRL_C_MSG,
                          SSH_BACKUP_ENCODING)
@@ -447,285 +448,24 @@ def create_containerapp(cmd,
                         user_assigned=None,
                         registry_identity=None,
                         workload_profile_name=None,
+                        termination_grace_period=None,
                         secret_volume_mount=None):
-    register_provider_if_needed(cmd, CONTAINER_APPS_RP)
-    validate_container_app_name(name, AppType.ContainerApp.name)
-    validate_create(registry_identity, registry_pass, registry_user, registry_server, no_wait)
-    validate_revision_suffix(revision_suffix)
+    raw_parameters = locals()
 
-    if registry_identity and not is_registry_msi_system(registry_identity):
-        logger.info("Creating an acrpull role assignment for the registry identity")
-        create_acrpull_role_assignment(cmd, registry_server, registry_identity, skip_error=True)
+    containerapp_create_decorator = ContainerAppCreateDecorator(
+        cmd=cmd,
+        client=ContainerAppClient,
+        raw_parameters=raw_parameters,
+        models="azext_containerapp._sdk_models"
+    )
+    containerapp_create_decorator.register_provider(CONTAINER_APPS_RP)
+    containerapp_create_decorator.validate_arguments()
 
-    if yaml:
-        if image or managed_env or min_replicas or max_replicas or target_port or ingress or\
-                revisions_mode or secrets or env_vars or cpu or memory or registry_server or\
-                registry_user or registry_pass or dapr_enabled or dapr_app_port or dapr_app_id or\
-                startup_command or args or tags:
-            not disable_warnings and logger.warning('Additional flags were passed along with --yaml. These flags will be ignored, and the configuration defined in the yaml will be used instead')
-        return create_containerapp_yaml(cmd=cmd, name=name, resource_group_name=resource_group_name, file_name=yaml, no_wait=no_wait)
-
-    if not image:
-        image = HELLO_WORLD_IMAGE
-
-    if managed_env is None:
-        raise RequiredArgumentMissingError('Usage error: --environment is required if not using --yaml')
-
-    # Validate managed environment
-    parsed_managed_env = parse_resource_id(managed_env)
-    managed_env_name = parsed_managed_env['name']
-    managed_env_rg = parsed_managed_env['resource_group']
-    managed_env_info = None
-
-    try:
-        managed_env_info = ManagedEnvironmentClient.show(cmd=cmd, resource_group_name=managed_env_rg, name=managed_env_name)
-    except:
-        pass
-
-    if not managed_env_info:
-        raise ValidationError("The environment '{}' does not exist. Specify a valid environment".format(managed_env))
-
-    location = managed_env_info["location"]
-    _ensure_location_allowed(cmd, location, CONTAINER_APPS_RP, "containerApps")
-
-    if not workload_profile_name and "workloadProfiles" in managed_env_info:
-        workload_profile_name = get_default_workload_profile_name_from_env(cmd, managed_env_info, managed_env_rg)
-
-    external_ingress = None
-    if ingress is not None:
-        if ingress.lower() == "internal":
-            external_ingress = False
-        elif ingress.lower() == "external":
-            external_ingress = True
-
-    ingress_def = None
-    if target_port is not None and ingress is not None:
-        ingress_def = IngressModel
-        ingress_def["external"] = external_ingress
-        ingress_def["targetPort"] = target_port
-        ingress_def["transport"] = transport
-        ingress_def["exposedPort"] = exposed_port if transport == "tcp" else None
-
-    secrets_def = None
-    if secrets is not None:
-        secrets_def = parse_secret_flags(secrets)
-
-    registries_def = None
-    if registry_server is not None and not is_registry_msi_system(registry_identity):
-        registries_def = RegistryCredentialsModel
-        registries_def["server"] = registry_server
-
-        # Infer credentials if not supplied and its azurecr
-        if (registry_user is None or registry_pass is None) and registry_identity is None:
-            registry_user, registry_pass = _infer_acr_credentials(cmd, registry_server, disable_warnings)
-
-        if not registry_identity:
-            registries_def["username"] = registry_user
-
-            if secrets_def is None:
-                secrets_def = []
-            registries_def["passwordSecretRef"] = store_as_secret_and_return_secret_ref(secrets_def, registry_user, registry_server, registry_pass, disable_warnings=disable_warnings)
-        else:
-            registries_def["identity"] = registry_identity
-
-    dapr_def = None
-    if dapr_enabled:
-        dapr_def = DaprModel
-        dapr_def["enabled"] = True
-        dapr_def["appId"] = dapr_app_id
-        dapr_def["appPort"] = dapr_app_port
-        dapr_def["appProtocol"] = dapr_app_protocol
-        dapr_def["httpReadBufferSize"] = dapr_http_read_buffer_size
-        dapr_def["httpMaxRequestSize"] = dapr_http_max_request_size
-        dapr_def["logLevel"] = dapr_log_level
-        dapr_def["enableApiLogging"] = dapr_enable_api_logging
-
-    service_def = None
-    if service_type:
-        service_def = ServiceModel
-        service_def["type"] = service_type
-
-    config_def = ConfigurationModel
-    config_def["secrets"] = secrets_def
-    config_def["activeRevisionsMode"] = revisions_mode
-    config_def["ingress"] = ingress_def
-    config_def["registries"] = [registries_def] if registries_def is not None else None
-    config_def["dapr"] = dapr_def
-    config_def["service"] = service_def if service_def is not None else None
-
-    # Identity actions
-    identity_def = ManagedServiceIdentityModel
-    identity_def["type"] = "None"
-
-    assign_system_identity = system_assigned
-    if user_assigned:
-        assign_user_identities = [x.lower() for x in user_assigned]
-    else:
-        assign_user_identities = []
-
-    if assign_system_identity and assign_user_identities:
-        identity_def["type"] = "SystemAssigned, UserAssigned"
-    elif assign_system_identity:
-        identity_def["type"] = "SystemAssigned"
-    elif assign_user_identities:
-        identity_def["type"] = "UserAssigned"
-
-    if assign_user_identities:
-        identity_def["userAssignedIdentities"] = {}
-        subscription_id = get_subscription_id(cmd.cli_ctx)
-
-        for r in assign_user_identities:
-            r = _ensure_identity_resource_id(subscription_id, resource_group_name, r)
-            identity_def["userAssignedIdentities"][r] = {}  # pylint: disable=unsupported-assignment-operation
-
-    scale_def = None
-    if min_replicas is not None or max_replicas is not None:
-        scale_def = ScaleModel
-        scale_def["minReplicas"] = min_replicas
-        scale_def["maxReplicas"] = max_replicas
-
-    if scale_rule_name:
-        if not scale_rule_type:
-            scale_rule_type = "http"
-        scale_rule_type = scale_rule_type.lower()
-        scale_rule_def = ScaleRuleModel
-        curr_metadata = {}
-        if scale_rule_http_concurrency:
-            if scale_rule_type in ('http', 'tcp'):
-                curr_metadata["concurrentRequests"] = str(scale_rule_http_concurrency)
-        metadata_def = parse_metadata_flags(scale_rule_metadata, curr_metadata)
-        auth_def = parse_auth_flags(scale_rule_auth)
-        if scale_rule_type == "http":
-            scale_rule_def["name"] = scale_rule_name
-            scale_rule_def["custom"] = None
-            scale_rule_def["http"] = {}
-            scale_rule_def["http"]["metadata"] = metadata_def
-            scale_rule_def["http"]["auth"] = auth_def
-        else:
-            scale_rule_def["name"] = scale_rule_name
-            scale_rule_def["http"] = None
-            scale_rule_def["custom"] = {}
-            scale_rule_def["custom"]["type"] = scale_rule_type
-            scale_rule_def["custom"]["metadata"] = metadata_def
-            scale_rule_def["custom"]["auth"] = auth_def
-        if not scale_def:
-            scale_def = ScaleModel
-        scale_def["rules"] = [scale_rule_def]
-
-    resources_def = None
-    if cpu is not None or memory is not None:
-        resources_def = ContainerResourcesModel
-        resources_def["cpu"] = cpu
-        resources_def["memory"] = memory
-
-    container_def = ContainerModel
-    container_def["name"] = container_name if container_name else name
-    container_def["image"] = image if not is_registry_msi_system(registry_identity) else HELLO_WORLD_IMAGE
-    if env_vars is not None:
-        container_def["env"] = parse_env_var_flags(env_vars)
-    if startup_command is not None:
-        container_def["command"] = startup_command
-    if args is not None:
-        container_def["args"] = args
-    if resources_def is not None:
-        container_def["resources"] = resources_def
-
-    template_def = TemplateModel
-
-    service_bindings_def_list = None
-    service_connectors_def_list = None
-
-    if service_bindings is not None:
-        service_connectors_def_list, service_bindings_def_list = parse_service_bindings(cmd, service_bindings,
-                                                                                        resource_group_name, name)
-        unique_bindings = check_unique_bindings(cmd, service_connectors_def_list, service_bindings_def_list,
-                                                resource_group_name, name)
-        if not unique_bindings:
-            raise ValidationError("Binding names across managed and dev services should be unique.")
-
-    template_def["containers"] = [container_def]
-    template_def["scale"] = scale_def
-    template_def["serviceBinds"] = service_bindings_def_list
-
-    if secret_volume_mount is not None:
-        volume_def = VolumeModel
-        volume_mount_def = VolumeMountModel
-        # generate a volume name
-        volume_def["name"] = _generate_secret_volume_name()
-        volume_def["storageType"] = "Secret"
-
-        # mount the volume to the container
-        volume_mount_def["volumeName"] = volume_def["name"]
-        volume_mount_def["mountPath"] = secret_volume_mount
-        container_def["volumeMounts"] = [volume_mount_def]
-        template_def["volumes"] = [volume_def]
-
-    if revision_suffix is not None and not is_registry_msi_system(registry_identity):
-        template_def["revisionSuffix"] = revision_suffix
-
-    containerapp_def = ContainerAppModel
-    containerapp_def["location"] = location
-    containerapp_def["identity"] = identity_def
-    containerapp_def["properties"]["environmentId"] = managed_env
-    containerapp_def["properties"]["configuration"] = config_def
-    containerapp_def["properties"]["template"] = template_def
-    containerapp_def["tags"] = tags
-
-    if workload_profile_name:
-        containerapp_def["properties"]["workloadProfileName"] = workload_profile_name
-        ensure_workload_profile_supported(cmd, managed_env_name, managed_env_rg, workload_profile_name, managed_env_info)
-
-    if registry_identity:
-        if is_registry_msi_system(registry_identity):
-            set_managed_identity(cmd, resource_group_name, containerapp_def, system_assigned=True)
-        else:
-            set_managed_identity(cmd, resource_group_name, containerapp_def, user_assigned=[registry_identity])
-    try:
-        r = ContainerAppClient.create_or_update(
-            cmd=cmd, resource_group_name=resource_group_name, name=name, container_app_envelope=containerapp_def, no_wait=no_wait)
-
-        if is_registry_msi_system(registry_identity):
-            while r["properties"]["provisioningState"] == "InProgress":
-                r = ContainerAppClient.show(cmd, resource_group_name, name)
-                time.sleep(10)
-            logger.info("Creating an acrpull role assignment for the system identity")
-            system_sp = r["identity"]["principalId"]
-            create_acrpull_role_assignment(cmd, registry_server, registry_identity=None, service_principal=system_sp)
-            container_def["image"] = image
-            safe_set(containerapp_def, "properties", "template", "revisionSuffix", value=revision_suffix)
-
-            registries_def = RegistryCredentialsModel
-            registries_def["server"] = registry_server
-            registries_def["identity"] = registry_identity
-            config_def["registries"] = [registries_def]
-
-            r = ContainerAppClient.create_or_update(cmd=cmd, resource_group_name=resource_group_name, name=name, container_app_envelope=containerapp_def, no_wait=no_wait)
-
-        if "properties" in r and "provisioningState" in r["properties"] and r["properties"]["provisioningState"].lower() == "waiting" and not no_wait:
-            not disable_warnings and logger.warning('Containerapp creation in progress. Please monitor the creation using `az containerapp show -n {} -g {}`'.format(name, resource_group_name))
-
-        if "configuration" in r["properties"] and "ingress" in r["properties"]["configuration"] and r["properties"]["configuration"]["ingress"] and "fqdn" in r["properties"]["configuration"]["ingress"]:
-            not disable_warnings and logger.warning("\nContainer app created. Access your app at https://{}/\n".format(r["properties"]["configuration"]["ingress"]["fqdn"]))
-        else:
-            target_port = target_port or "<port>"
-            not disable_warnings and logger.warning("\nContainer app created. To access it over HTTPS, enable ingress: "
-                                                    "az containerapp ingress enable -n %s -g %s --type external --target-port %s"
-                                                    " --transport auto\n", name, resource_group_name, target_port)
-
-        if service_connectors_def_list is not None:
-            linker_client = get_linker_client(cmd)
-
-            for item in service_connectors_def_list:
-                while r["properties"]["provisioningState"].lower() == "inprogress":
-                    r = ContainerAppClient.show(cmd, resource_group_name, name)
-                    time.sleep(1)
-                linker_client.linker.begin_create_or_update(resource_uri=r["id"],
-                                                            parameters=item["parameters"],
-                                                            linker_name=item["linker_name"]).result()
-
-        return r
-    except Exception as e:
-        handle_raw_exception(e)
+    containerapp_def = containerapp_create_decorator.construct_containerapp()
+    r = containerapp_create_decorator.create_containerapp(containerapp_def)
+    containerapp_def = containerapp_create_decorator.construct_containerapp_for_post_process(containerapp_def, r)
+    r = containerapp_create_decorator.post_process_containerapp(containerapp_def, r)
+    return r
 
 
 def update_containerapp_logic(cmd,
@@ -758,6 +498,7 @@ def update_containerapp_logic(cmd,
                               ingress=None,
                               target_port=None,
                               workload_profile_name=None,
+                              termination_grace_period=None,
                               registry_server=None,
                               registry_user=None,
                               registry_pass=None,
@@ -875,6 +616,9 @@ def update_containerapp_logic(cmd,
     if revision_suffix is not None:
         new_containerapp["properties"]["template"] = {} if "template" not in new_containerapp["properties"] else new_containerapp["properties"]["template"]
         new_containerapp["properties"]["template"]["revisionSuffix"] = revision_suffix
+
+    if termination_grace_period is not None:
+        safe_set(new_containerapp, "properties", "template", "terminationGracePeriodSeconds", value=termination_grace_period)
 
     if workload_profile_name:
         new_containerapp["properties"]["workloadProfileName"] = workload_profile_name
@@ -1215,6 +959,7 @@ def update_containerapp(cmd,
                         args=None,
                         tags=None,
                         workload_profile_name=None,
+                        termination_grace_period=None,
                         no_wait=False,
                         secret_volume_mount=None):
     _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
@@ -1245,52 +990,48 @@ def update_containerapp(cmd,
                                      args=args,
                                      tags=tags,
                                      workload_profile_name=workload_profile_name,
+                                     termination_grace_period=termination_grace_period,
                                      no_wait=no_wait,
                                      secret_volume_mount=secret_volume_mount)
 
 
 def show_containerapp(cmd, name, resource_group_name, show_secrets=False):
-    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+    raw_parameters = locals()
+    containerapp_base_decorator = BaseContainerAppDecorator(
+        cmd=cmd,
+        client=ContainerAppClient,
+        raw_parameters=raw_parameters,
+        models="azext_containerapp._sdk_models"
+    )
+    containerapp_base_decorator.validate_subscription_registered(CONTAINER_APPS_RP)
 
-    try:
-        r = ContainerAppClient.show(cmd=cmd, resource_group_name=resource_group_name, name=name)
-        if show_secrets:
-            _get_existing_secrets(cmd, resource_group_name, name, r)
-        return r
-    except CLIError as e:
-        handle_raw_exception(e)
+    return containerapp_base_decorator.show_containerapp()
 
 
 def list_containerapp(cmd, resource_group_name=None, managed_env=None):
-    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+    raw_parameters = locals()
+    containerapp_base_decorator = BaseContainerAppDecorator(
+        cmd=cmd,
+        client=ContainerAppClient,
+        raw_parameters=raw_parameters,
+        models="azext_containerapp._sdk_models"
+    )
+    containerapp_base_decorator.validate_subscription_registered(CONTAINER_APPS_RP)
 
-    try:
-        containerapps = []
-        if resource_group_name is None:
-            containerapps = ContainerAppClient.list_by_subscription(cmd=cmd)
-        else:
-            containerapps = ContainerAppClient.list_by_resource_group(cmd=cmd, resource_group_name=resource_group_name)
-
-        if managed_env:
-            env_name = parse_resource_id(managed_env)["name"].lower()
-            if "resource_group" in parse_resource_id(managed_env):
-                ManagedEnvironmentClient.show(cmd, parse_resource_id(managed_env)["resource_group"], parse_resource_id(managed_env)["name"])
-                containerapps = [c for c in containerapps if c["properties"]["environmentId"].lower() == managed_env.lower()]
-            else:
-                containerapps = [c for c in containerapps if parse_resource_id(c["properties"]["environmentId"])["name"].lower() == env_name]
-
-        return containerapps
-    except CLIError as e:
-        handle_raw_exception(e)
+    return containerapp_base_decorator.list_containerapp()
 
 
 def delete_containerapp(cmd, name, resource_group_name, no_wait=False):
-    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+    raw_parameters = locals()
+    containerapp_base_decorator = BaseContainerAppDecorator(
+        cmd=cmd,
+        client=ContainerAppClient,
+        raw_parameters=raw_parameters,
+        models="azext_containerapp._sdk_models"
+    )
+    containerapp_base_decorator.validate_subscription_registered(CONTAINER_APPS_RP)
 
-    try:
-        return ContainerAppClient.delete(cmd=cmd, name=name, resource_group_name=resource_group_name, no_wait=no_wait)
-    except CLIError as e:
-        handle_raw_exception(e)
+    return containerapp_base_decorator.delete_containerapp()
 
 
 def create_managed_environment(cmd,
@@ -1314,6 +1055,7 @@ def create_managed_environment(cmd,
                                certificate_file=None,
                                certificate_password=None,
                                enable_workload_profiles=False,
+                               mtls_enabled=None,
                                no_wait=False):
     if zone_redundant:
         if not infrastructure_subnet_resource_id:
@@ -1390,6 +1132,8 @@ def create_managed_environment(cmd,
             raise ValidationError('Infrastructure subnet resource ID needs to be supplied for internal only environments.')
         managed_env_def["properties"]["vnetConfiguration"]["internal"] = True
 
+    if mtls_enabled is not None:
+        safe_set(managed_env_def, "properties", "peerAuthentication", "mtls", "enabled", value=mtls_enabled)
     try:
         r = ManagedEnvironmentClient.create(
             cmd=cmd, resource_group_name=resource_group_name, name=name, managed_environment_envelope=managed_env_def, no_wait=no_wait)
@@ -1424,6 +1168,7 @@ def update_managed_environment(cmd,
                                workload_profile_name=None,
                                min_nodes=None,
                                max_nodes=None,
+                               mtls_enabled=None,
                                no_wait=False):
     if logs_destination == "log-analytics" or logs_customer_id or logs_key:
         if logs_destination != "log-analytics":
@@ -1493,6 +1238,9 @@ def update_managed_environment(cmd,
             workload_profiles[idx] = profile
 
         safe_set(env_def, "properties", "workloadProfiles", value=workload_profiles)
+
+    if mtls_enabled is not None:
+        safe_set(env_def, "properties", "peerAuthentication", "mtls", "enabled", value=mtls_enabled)
 
     try:
         r = ManagedEnvironmentClient.update(
@@ -2753,6 +2501,7 @@ def create_or_update_github_action(cmd,
                                    service_principal_client_id=None,
                                    service_principal_client_secret=None,
                                    service_principal_tenant_id=None,
+                                   trigger_existing_workflow=False,
                                    no_wait=False):
     if not token and not login_with_github:
         raise_missing_token_suggestion()
@@ -2776,6 +2525,16 @@ def create_or_update_github_action(cmd,
         if not service_principal_client_id or not service_principal_client_secret or not service_principal_tenant_id:
             raise RequiredArgumentMissingError('Service principal client ID, secret and tenant ID are required to add github actions for the first time. Please create one using the command \"az ad sp create-for-rbac --name {{name}} --role contributor --scopes /subscriptions/{{subscription}}/resourceGroups/{{resourceGroup}} --sdk-auth\"') from ex
         source_control_info = SourceControlModel
+
+    # Need to trigger the workflow manually if it already exists (performing an update)
+    try:
+        workflow_name = GitHubActionClient.get_workflow_name(cmd=cmd, repo=repo, branch_name=branch, container_app_name=name, token=token)
+        if workflow_name is not None:
+            if trigger_existing_workflow:
+                trigger_workflow(token, repo, workflow_name, branch)
+            return source_control_info
+    except:  # pylint: disable=bare-except
+        pass
 
     source_control_info["properties"]["repoUrl"] = repo_url
     source_control_info["properties"]["branch"] = branch
@@ -2822,7 +2581,20 @@ def create_or_update_github_action(cmd,
         logger.warning("Creating Github action...")
         r = GitHubActionClient.create_or_update(cmd=cmd, resource_group_name=resource_group_name, name=name, github_action_envelope=source_control_info, headers=headers, no_wait=no_wait)
         if not no_wait:
-            await_github_action(token, repo, r["properties"]["githubActionConfiguration"]["workflowName"])
+            WORKFLOW_POLL_RETRY = 3
+            WORKFLOW_POLL_SLEEP = 10
+
+            # Poll for the workflow file just created (may take up to 30s)
+            for _ in range(0, WORKFLOW_POLL_RETRY):
+                time.sleep(WORKFLOW_POLL_SLEEP)
+                workflow_name = GitHubActionClient.get_workflow_name(cmd=cmd, repo=repo, branch_name=branch, container_app_name=name, token=token)
+                if workflow_name is not None:
+                    await_github_action(token, repo, workflow_name)
+                    return r
+
+            raise ValidationError(
+                "Failed to find workflow file for Container App '{}' in .github/workflow folder for repo '{}'. ".format(name, repo) +
+                "If this file was removed, please use the 'az containerapp github-action delete' command to disconnect the removed workflow file connection.")
         return r
     except Exception as e:
         handle_raw_exception(e)
@@ -3521,6 +3293,120 @@ def show_ingress_sticky_session(cmd, name, resource_group_name):
         raise ValidationError("Ingress must be enabled to enable sticky sessions. Try running `az containerapp ingress -h` for more info.") from e
 
 
+def enable_cors_policy(cmd, name, resource_group_name, allowed_origins, allowed_methods=None, allowed_headers=None, expose_headers=None, allow_credentials=None, max_age=None, no_wait=False):
+    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+
+    if not allowed_origins:
+        raise ValidationError("Allowed origins must be specified.")
+
+    containerapp_def = None
+    try:
+        containerapp_def = ContainerAppClient.show(cmd=cmd, resource_group_name=resource_group_name, name=name)
+    except:
+        pass
+
+    if not containerapp_def:
+        raise ResourceNotFoundError(f"The containerapp '{name}' does not exist in group '{resource_group_name}'")
+
+    containerapp_patch = {}
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowedOrigins", value=allowed_origins)
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowedMethods", value=allowed_methods)
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowedHeaders", value=allowed_headers)
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "exposeHeaders", value=expose_headers)
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowCredentials", value=allow_credentials)
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "maxAge", value=max_age)
+    try:
+        r = ContainerAppClient.update(
+            cmd=cmd, resource_group_name=resource_group_name, name=name, container_app_envelope=containerapp_patch, no_wait=no_wait)
+        return safe_get(r, "properties", "configuration", "ingress", "corsPolicy", default={})
+    except Exception as e:
+        handle_raw_exception(e)
+
+
+def disable_cors_policy(cmd, name, resource_group_name, no_wait=False):
+    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+
+    containerapp_def = None
+    try:
+        containerapp_def = ContainerAppClient.show(cmd=cmd, resource_group_name=resource_group_name, name=name)
+    except:
+        pass
+
+    if not containerapp_def:
+        raise ResourceNotFoundError(f"The containerapp '{name}' does not exist in group '{resource_group_name}'")
+
+    containerapp_patch = {}
+    safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", value=None)
+    try:
+        r = ContainerAppClient.update(
+            cmd=cmd, resource_group_name=resource_group_name, name=name, container_app_envelope=containerapp_patch, no_wait=no_wait)
+        return safe_get(r, "properties", "configuration", "ingress", default={})
+    except Exception as e:
+        handle_raw_exception(e)
+
+
+def update_cors_policy(cmd, name, resource_group_name, allowed_origins=None, allowed_methods=None, allowed_headers=None, expose_headers=None, allow_credentials=None, max_age=None, no_wait=False):
+    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+
+    containerapp_def = None
+    try:
+        containerapp_def = ContainerAppClient.show(cmd=cmd, resource_group_name=resource_group_name, name=name)
+    except:
+        pass
+
+    if not containerapp_def:
+        raise ResourceNotFoundError(f"The containerapp '{name}' does not exist in group '{resource_group_name}'")
+
+    if allowed_origins is not None and len(allowed_origins) == 0:
+        raise RequiredArgumentMissingError("allowed-origins must be specified if provided.")
+
+    reset_max_age = False
+    if max_age == "":
+        reset_max_age = True
+
+    containerapp_patch = {}
+    if allowed_origins is not None:
+        safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowedOrigins", value=allowed_origins)
+    if allowed_methods is not None:
+        safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowedMethods", value=allowed_methods)
+    if allowed_headers is not None:
+        safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowedHeaders", value=allowed_headers)
+    if expose_headers is not None:
+        safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "exposeHeaders", value=expose_headers)
+    if allow_credentials is not None:
+        safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "allowCredentials", value=allow_credentials)
+    if max_age is not None:
+        if reset_max_age:
+            safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "maxAge", value=None)
+        else:
+            safe_set(containerapp_patch, "properties", "configuration", "ingress", "corsPolicy", "maxAge", value=max_age)
+
+    try:
+        r = ContainerAppClient.update(
+            cmd=cmd, resource_group_name=resource_group_name, name=name, container_app_envelope=containerapp_patch, no_wait=no_wait)
+        return safe_get(r, "properties", "configuration", "ingress", "corsPolicy", default={})
+    except Exception as e:
+        handle_raw_exception(e)
+
+
+def show_cors_policy(cmd, name, resource_group_name):
+    _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
+
+    containerapp_def = None
+    try:
+        containerapp_def = ContainerAppClient.show(cmd=cmd, resource_group_name=resource_group_name, name=name)
+    except:
+        pass
+
+    if not containerapp_def:
+        raise ResourceNotFoundError(f"The containerapp '{name}' does not exist in group '{resource_group_name}'")
+
+    try:
+        return safe_get(containerapp_def, "properties", "configuration", "ingress", "corsPolicy", default={})
+    except Exception as e:
+        raise ValidationError("CORS must be enabled to enable CORS policy. Try running `az containerapp ingress cors enable -h` for more info.") from e
+
+
 def show_registry(cmd, name, resource_group_name, server):
     _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
 
@@ -4194,13 +4080,16 @@ def stream_environment_logs(cmd, resource_group_name, name, follow=False, tail=N
             raise ValidationError("--tail must be between 0 and 300.")
 
     env = show_managed_environment(cmd, name, resource_group_name)
-    sub = get_subscription_id(cmd.cli_ctx)
+    url = safe_get(env, "properties", "eventStreamEndpoint")
+
+    if url is None:
+        sub = get_subscription_id(cmd.cli_ctx)
+        base_url = f"https://{format_location(env['location'])}.azurecontainerapps.dev"
+        url = (f"{base_url}/subscriptions/{sub}/resourceGroups/{resource_group_name}/managedEnvironments/{name}"
+               f"/eventstream")
+
     token_response = ManagedEnvironmentClient.get_auth_token(cmd, resource_group_name, name)
     token = token_response["properties"]["token"]
-    base_url = f"https://{env['location']}.azurecontainerapps.dev"
-
-    url = (f"{base_url}/subscriptions/{sub}/resourceGroups/{resource_group_name}/managedEnvironments/{name}"
-           f"/eventstream")
 
     logger.info("connecting to : %s", url)
     request_params = {"follow": str(follow).lower(),